Fix various warnings
Found while building on Debian 7

Github-Pull: #6133
Rebased-From: e617fe25785bfba30bcb0a8d8af6797b44ba07f5

+++ src/txmempool.cpp
@@ -499,7 +499,7 @@ void CTxMemPool::removeForReorg(const CCoinsViewCache *pcoins, unsigned int nMem
                     continue;
                 const CCoins *coins = pcoins->AccessCoins(txin.prevout.hash);
                 if (fSanityCheck) assert(coins);
-                if (!coins || (coins->IsCoinBase() && nMemPoolHeight - coins->nHeight < COINBASE_MATURITY)) {
+                if (!coins || (coins->IsCoinBase() && ((signed long)nMemPoolHeight) - coins->nHeight < COINBASE_MATURITY)) {
                     transactionsToRemove.push_back(tx);
                     break;
                 }
